feat: add exceptions to CustomerService

+++ src/main/java/com/vodafone/service/CustomerService.java
@@ -1,20 +1,24 @@
 package com.vodafone.service;
 
+import com.vodafone.exception.IncompleteUserAttributesException;
+import com.vodafone.exception.NullCustomerException;
+import com.vodafone.exception.NullIdException;
 import com.vodafone.model.Customer;
 import com.vodafone.model.Role;
 import com.vodafone.model.UserStatus;
 import com.vodafone.repository.customer.CustomerRepository;
+import com.vodafone.repository.customer.ICustomerRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class CustomerService {
-    private final CustomerRepository customerRepository;
+    private final ICustomerRepository customerRepository;
 
     private final HashService hashService;
 
-    public CustomerService(CustomerRepository customerRepository, HashService hashService) {
+    public CustomerService(ICustomerRepository customerRepository, HashService hashService) {
         this.customerRepository = customerRepository;
         this.hashService = hashService;
         /*Customer customer = new Customer();
@@ -28,14 +32,27 @@ public class CustomerService {
          */
     }
 
+    public boolean isCustomerAttributesValid(Customer customer) {
+        return customer.getUserName() != null && customer.getEmail() != null
+                && customer.getPassword() != null && customer.getRole() != null
+                && customer.getUserStatus() != null;
+    }
 
     public boolean create(Customer customer) {
+        if (customer == null)
+            throw new NullCustomerException("Null customer entity is provided");
+        if (!isCustomerAttributesValid(customer))
+            throw new IncompleteUserAttributesException("Customer Data is not completed, nulls exist");
         String originalPassword = customer.getPassword();
         customer.setPassword(hashService.encryptPassword(originalPassword, customer.getEmail()));
         return customerRepository.create(customer);
     }
 
     public boolean update(Long id, Customer updatedCustomer) {
+        if (id == null)
+            throw new NullIdException("Null customer id is provided");
+        if (updatedCustomer == null)
+            throw new NullCustomerException("Null updated customer entity is provided");
         return customerRepository.update(id, updatedCustomer);
     }
 
@@ -44,18 +61,26 @@ public class CustomerService {
     }
 
     public boolean delete(Long id) {
+        if (id == null)
+            throw new NullIdException("Null customer id is provided");
         return customerRepository.delete(id);
     }
 
     public Customer get(Long id) {
+        if (id == null)
+            throw new NullIdException("Null customer id is provided");
         return customerRepository.get(id);
     }
 
     public Customer getByMail(String email) {
+        if (email == null)
+            throw new NullPointerException("Null Email is provided");
         return customerRepository.getByMail(email);
     }
 
     public Customer getByUserName(String username) {
+        if (username == null)
+            throw new NullPointerException("Null Username is provided");
         return customerRepository.getByUserName(username);
     }
 
@@ -64,10 +89,16 @@ public class CustomerService {
     }
 
     public boolean resetPassword(String email, String password) {
+        if (email == null)
+            throw new NullPointerException("Null Email is provided");
+        if (password == null)
+            throw new NullPointerException("Null Password is provided");
         return customerRepository.resetPassword(email, password);
     }
 
     public boolean expireOtp(String userName) {
+        if (userName == null)
+            throw new NullPointerException("Null Username is provided");
         return this.customerRepository.expireOtp(userName);
     }
